Split and comment on hidden dependency
